Ai now has a shooting range and a sight range.
If the player is outside of the AI's shooting range but within the sight range the AI will move towards you but not shoot until you are within shooting range.

If the player is within shooting range but outside the stopping distance the AI will shoot while moving towards you but will still stop when reaching the stopping distance.